requirements.txt
flask
pypdf
openai
python-dotenv
requests